+++ packages/address/CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Added
+
+- Add `AddressFormatter.resetCache()` to be able to avoid caching side-effects [[#2179](https://github.com/Shopify/quilt/pull/2179)]
+
+### Changed
+
+- `getCountries` now checks the cache before fetching. Use the static `resetCache` method if you'd like to force a refetch. [[#2179](https://github.com/Shopify/quilt/pull/2179)]
 
 ## 3.1.1 - 2022-03-09